+++ docs/operators-and-hooks-ref.rst
@@ -352,7 +352,7 @@ These integrations allow you to perform various operations within the Amazon Web
    * - `Amazon EMR <https://aws.amazon.com/emr/>`__
      -
      - :mod:`airflow.providers.amazon.aws.hooks.emr`
-     - :mod:`airflow.contrib.operators.emr_add_steps_operator`,
+     - :mod:`airflow.providers.amazon.aws.operators.emr_add_steps`,
        :mod:`airflow.contrib.operators.emr_create_job_flow_operator`,
        :mod:`airflow.contrib.operators.emr_terminate_job_flow_operator`
      - :mod:`airflow.contrib.sensors.emr_base_sensor`,
