+++ tests/providers/amazon/aws/operators/test_emr_add_steps.py
@@ -22,9 +22,9 @@ from datetime import timedelta
 from unittest.mock import MagicMock, patch
 
 from airflow import DAG
-from airflow.contrib.operators.emr_add_steps_operator import EmrAddStepsOperator
 from airflow.exceptions import AirflowException
 from airflow.models import TaskInstance
+from airflow.providers.amazon.aws.operators.emr_add_steps import EmrAddStepsOperator
 from airflow.utils import timezone
 
 DEFAULT_DATE = timezone.datetime(2017, 1, 1)
